[WIP] Pid refactoring

+++ plugins/controlboard/src/ControlBoardDriverImpedancePositionControl.cpp
@@ -13,7 +13,7 @@ using namespace yarp::dev;
  */
 bool GazeboYarpControlBoardDriver::getImpedance(int j, double *stiffness, double *damping)
 {
-    if (j >= 0 && j < m_numberOfJoints) {
+    if (j >= 0 && static_cast<size_t>(j) < m_numberOfJoints) {
         *stiffness = m_impedancePosPDs[j].p;
         *damping = m_impedancePosPDs[j].d;
         return true;
@@ -26,7 +26,7 @@ bool GazeboYarpControlBoardDriver::getImpedance(int j, double *stiffness, double
  */
 bool GazeboYarpControlBoardDriver::setImpedance(int j, double stiffness, double damping)
 {
-    if (j >= 0 && j < m_numberOfJoints) {
+    if (j >= 0 && static_cast<size_t>(j) < m_numberOfJoints) {
         m_impedancePosPDs[j].p = stiffness;
         m_impedancePosPDs[j].d = damping;
         return true;
@@ -39,7 +39,7 @@ bool GazeboYarpControlBoardDriver::setImpedance(int j, double stiffness, double 
  */
 bool GazeboYarpControlBoardDriver::setImpedanceOffset(int j, double offset)
 {
-    if(j >= 0 && j < m_numberOfJoints)
+    if (j >= 0 && static_cast<size_t>(j) < m_numberOfJoints)
     {
         m_torqueOffset[j] = offset;
         return true;
@@ -52,7 +52,7 @@ bool GazeboYarpControlBoardDriver::setImpedanceOffset(int j, double offset)
  */
 bool GazeboYarpControlBoardDriver::getImpedanceOffset(int j, double* offset)
 {
-    if(j >= 0 && j < m_numberOfJoints)
+    if (j >= 0 && static_cast<size_t>(j) < m_numberOfJoints)
     {
         *offset = m_torqueOffset[j];
         return true;
@@ -65,7 +65,7 @@ bool GazeboYarpControlBoardDriver::getImpedanceOffset(int j, double* offset)
  */
 bool GazeboYarpControlBoardDriver::getCurrentImpedanceLimit(int j, double *min_stiff, double *max_stiff, double *min_damp, double *max_damp)
 {
-    if(j >= 0 && j < m_numberOfJoints)
+    if (j >= 0 && static_cast<size_t>(j) < m_numberOfJoints)
     {
         //Hardcoded numbers...just to try
         *min_stiff = m_minStiffness[j];
